feat: Add functionality to generate weekly activated codes report
This commit adds the functionality to generate a weekly activated codes report. It includes a new command "Faollashtirilgan kodlar hisoboti" that can be triggered by admins. The command retrieves the total number of activated codes and the number of codes activated during the latest game week. The report is then sent as a message to the admin. This feature enhances the admin's ability to track code activation and monitor the progress of the game.